accel/amdxdna: Consolidate common driver code into amdxdna_drv
Extract common DRM and device management code from PCI-specific driver
into new amdxdna_drv.c/h files. This enables code sharing between PCI
and auxiliary bus driver variants.

The refactoring moves the following to common code:
- DRM driver operations (open, close, file_ops)
- IOCTL handlers (get_info, set_state, get_array)
- Device initialization and cleanup (amdxdna_dev_init/cleanup)
- Client management (amdxdna_client_cleanup)
- Common data structures and definitions

PCI driver retains:
- Hardware-specific probe/remove functions
- PCI device table and identification
- IOMMU initialization (PCI-specific)
- SR-IOV configuration
- Power management operations

Changes:
- Add amdxdna_drv.c with common DRM/device functions
- Add amdxdna_drv.h with shared structures and ops definitions
- Simplify amdxdna_pci_drv.c to use common init/cleanup functions
- Reduce amdxdna_pci_drv.h to PCI-specific definitions only
- Update amdxdna_aux_drv.c to use common module definitions
- Move struct amdxdna_dev_info to PCI header (hardware-specific)
- Add amdxdna_for_each_client macro to common header
- Update Kbuild to include amdxdna_drv.o

This refactoring eliminates ~460 lines of duplicate code while
maintaining full compatibility with existing PCI driver functionality.
