+++ drivers/accel/amdxdna/Kbuild
@@ -23,6 +23,7 @@ ccflags-${UMQ_HELLO_TEST} += -DUMQ_HELLO_TEST
 obj-m	+= amdxdna.o
 # Only objects that have corresponding .c sources in this directory
 amdxdna-y := \
+	amdxdna_drv.o \
 	amdxdna_solver.o \
 	amdxdna_ctx.o \
 	amdxdna_gem.o \
