+++ drivers/accel/amdxdna/amdxdna_ctx.c
@@ -16,8 +16,9 @@
 
 #include "amdxdna_ctx.h"
 #include "amdxdna_gem.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 #include "amdxdna_pm.h"
+#include "amdxdna_solver.h"
 
 #define MAX_HWCTX_ID		255
 #define MAX_ARG_COUNT		4095
@@ -461,7 +462,6 @@ put_shmem_bo:
 void amdxdna_sched_job_cleanup(struct amdxdna_sched_job *job)
 {
 	trace_amdxdna_debug_point(job->hwctx->name, job->seq, "job release");
-	amdxdna_pm_suspend_put(job->hwctx->client->xdna);
 	amdxdna_arg_bos_put(job);
 	amdxdna_gem_put_obj(job->cmd_bo);
 	dma_fence_put(job->fence);
@@ -497,12 +497,6 @@ int amdxdna_cmd_submit(struct amdxdna_client *client,
 	if (ret) {
 		XDNA_ERR(xdna, "Argument BOs lookup failed, ret %d", ret);
 		goto cmd_put;
-	}
-
-	ret = amdxdna_pm_resume_get(xdna);
-	if (ret) {
-		XDNA_ERR(xdna, "Resume failed, ret %d", ret);
-		goto put_bos;
 	}
 
 	idx = srcu_read_lock(&client->hwctx_srcu);
@@ -545,8 +539,6 @@ put_fence:
 	dma_fence_put(job->fence);
 unlock_srcu:
 	srcu_read_unlock(&client->hwctx_srcu, idx);
-	amdxdna_pm_suspend_put(xdna);
-put_bos:
 	amdxdna_arg_bos_put(job);
 cmd_put:
 	amdxdna_gem_put_obj(job->cmd_bo);
